se cambian los uint8_t por uint16_t y se crean los metodos enviar_dos_bytes y recibir_dos_bytes

+++ common/common_bala.h
@@ -5,15 +5,15 @@
 
 class Bala: public Entidad {
 protected:
-    uint8_t pos_x_final;
-    uint8_t pos_y_final;
+    uint16_t pos_x_final;
+    uint16_t pos_y_final;
     uint8_t direccion;
 
 public:
-    Bala(uint8_t id, uint8_t pos_x, uint8_t pos_y, uint8_t pos_x_final, uint8_t pos_y_final,
+    Bala(uint16_t id, uint16_t pos_x, uint16_t pos_y, uint16_t pos_x_final, uint16_t pos_y_final,
          uint8_t direccion);
-    uint8_t get_pos_x_final();
-    uint8_t get_pos_y_final();
+    uint16_t get_pos_x_final();
+    uint16_t get_pos_y_final();
     uint8_t get_direccion();
 };
