+++ common/common_caja.h
@@ -5,12 +5,12 @@
 
 class Caja: public Entidad {
 private:
-    uint8_t tipo;
-    uint8_t recompensa;
+    uint16_t tipo;
+    uint16_t recompensa;
     bool esta_vacia;
 
 public:
-    Caja(uint8_t id, uint8_t pos_x, uint8_t pos_y, uint8_t recompensa);
+    Caja(uint16_t id, uint16_t pos_x, uint16_t pos_y, uint8_t recompensa);
     uint8_t get_recompensa();
 };
